refactor: extract round-scoring & round-stats cluster from state.py (#1271)
Pull the round-end scoring pass + round-stats/highlights recording out of
the state.py God-Object into a new RoundScoringMixin (state_scoring.py).
Stacked on the vote-window cut (#1339).

Moved (pure code-move, behavior-preserving):
- _score_round            — scoring pass + closest-wins + round_results
- _record_round_stats     — highlights, analytics, song-result stats
- _record_round_highlights — per-round highlight detection

_score_all_players deliberately STAYS on GameState: the vote-window path
(_finalize_title_artist_window) reuses self._score_all_players so the
round-end and deferred-rescore paths cannot drift (one source of truth).
The REVEAL announcement / auto-advance / party-light helpers also stay —
they couple to TTS / vote-window / lights, out of scope for a scoring cut.

state.py 2174 -> 1955 lines (-219). GameState now mixes in 7 mixins.

Tests: 885 passed, 2 skipped. ruff check + ruff format --check both clean.
